+++ safespaces/devices/vive.lua
@@ -1,0 +1,9 @@
+return {
+	display = '^HTCVIVE',
+	oversample_w = 1.6,
+	oversample_h = 1.6,
+	distortion_model = "basic",
+	prelaunch = true,
+	bindings = {
+	}
+};