+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 0.2 (ongoing)
 
-* Added 'prelaunch' option to device profiles
+* Added 'prelaunch' option to device profile
 * Added Rift device profile
+* Added Vive device profile
+* Added menu controls for grab toggle
 
 0.1 : 28/3/2018 - First Release